fix(stdlib): quote finite floats as lua hex literals

+++ crates/elara-test/tests/conformance.rs
@@ -729,14 +729,14 @@ fn conformance_standard_library_fixtures() {
     assert_success_fixture(
         "stdlib/string_format_quote_scalars.lua",
         vec![
-            Value::integer(15),
+            Value::integer(20),
             Value::integer(110),
             Value::integer(58),
             Value::integer(116),
             Value::integer(58),
             Value::integer(45),
             Value::integer(58),
-            Value::integer(53),
+            Value::integer(49),
         ],
     );
     assert_success_fixture(
